fixing multi review
