+++ api/src/main/java/org/pmiops/workbench/cohortbuilder/CohortBuilderService.java
@@ -5,23 +5,7 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 import org.pmiops.workbench.db.model.DbConceptSetConceptId;
-import org.pmiops.workbench.model.AgeType;
-import org.pmiops.workbench.model.AgeTypeCount;
-import org.pmiops.workbench.model.CardCount;
-import org.pmiops.workbench.model.Criteria;
-import org.pmiops.workbench.model.CriteriaAttribute;
-import org.pmiops.workbench.model.CriteriaListWithCountResponse;
-import org.pmiops.workbench.model.CriteriaMenu;
-import org.pmiops.workbench.model.CriteriaType;
-import org.pmiops.workbench.model.DataFilter;
-import org.pmiops.workbench.model.DemoChartInfo;
-import org.pmiops.workbench.model.DomainCard;
-import org.pmiops.workbench.model.EthnicityInfo;
-import org.pmiops.workbench.model.GenderOrSexType;
-import org.pmiops.workbench.model.ParticipantDemographics;
-import org.pmiops.workbench.model.SearchRequest;
-import org.pmiops.workbench.model.SurveyModule;
-import org.pmiops.workbench.model.SurveyVersion;
+import org.pmiops.workbench.model.*;
 
 public interface CohortBuilderService {
 
@@ -55,6 +39,8 @@ public interface CohortBuilderService {
   Long countParticipants(SearchRequest request);
 
   List<AgeTypeCount> findAgeTypeCounts();
+
+  List<CohortChartData> findCohortChartData(SearchRequest searchRequest, Domain domain, int limit);
 
   List<CriteriaAttribute> findCriteriaAttributeByConceptId(Long conceptId);
